issues: move all issues components into subdir.
